+++ contracts/allocators/AllocatorWithManager.sol
@@ -48,7 +48,7 @@ contract AllocatorWithManager is
         address airnode,
         uint256 slotIndex,
         bytes32 subscriptionId,
-        uint64 expirationTimestamp
+        uint32 expirationTimestamp
     ) external override {
         require(
             hasSlotSetterRoleOrIsManager(_msgSender()),
